requirements.txt
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
asttokens==3.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.7.0
jupyter_core==5.9.1
langchain-core==1.2.6
langchain-text-splitters==1.1.0
langsmith==0.6.0
matplotlib-inline==0.2.1
-e git+https://github.com/andreasrieger/on-prem-llms.git@b278f71a32c4157851c37bf14788b0fdccd03029#egg=mypackage
nest-asyncio==1.6.0
numpy==2.4.0
ollama==0.6.1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.2
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
wcwidth==0.2.14
zstandard==0.25.0
